# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: MathGenie API
  description: Server для обучающего приложения MathGenie
  version: 1.0.0
  contact:
    name: MathGenie Support
    email: asoltobekovv@gmail.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.mathgenie.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Schedule:
      type: object
      properties:
        _id:
          type: string
          description: Уникальный идентификатор занятия
        date:
          type: string
          format: date
          description: Дата проведения занятия
        startTime:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
          description: Время начала занятия (формат HH:MM)
        endTime:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
          description: Время окончания занятия (формат HH:MM)
        title:
          type: string
          description: Тема урока
        description:
          type: string
          description: Краткое описание занятия
        format:
          type: string
          enum: [онлайн, оффлайн, запись]
          default: онлайн
          description: Формат занятия
        status:
          type: string
          enum: [запланирован, проведён, перенесён, отменён]
          default: запланирован
          description: Статус занятия
        teacher:
          type: string
          description: Имя преподавателя
        materials:
          type: array
          items:
            type: string
          description: Ссылки на материалы (PDF, видео, статьи)
        streamLink:
          type: string
          description: Ссылка на трансляцию (если есть)
        homework:
          type: string
          description: Описание домашнего задания
        homeworkDeadline:
          type: string
          format: date
          description: Дата сдачи домашнего задания
        createdAt:
          type: string
          format: date-time
          description: Дата создания записи
        updatedAt:
          type: string
          format: date-time
          description: Дата последнего обновления
paths:
...
  /schedule:
    get:
      summary: Получить расписание занятий
      description: |
        Возвращает список всех занятий с возможностью фильтрации по дате, статусу, 
        формату и преподавателю. Студенты могут просматривать расписание.
      tags:
        - Schedule
      parameters:
        - name: date
          in: query
          description: Фильтр по дате (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
        - name: status
          in: query
          description: Фильтр по статусу занятия
          required: false
          schema:
            type: string
            enum: [запланирован, проведён, перенесён, отменён]
        - name: format
          in: query
          description: Фильтр по формату занятия
          required: false
          schema:
            type: string
            enum: [онлайн, оффлайн, запись]
        - name: teacher
          in: query
          description: Фильтр по имени преподавателя
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Успешно получено расписание
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Schedule'
        '500':
          description: Ошибка сервера

    post:
      summary: Создать новое занятие
      description: |
        Создаёт новое занятие в расписании. Доступно только администраторам.
        Автоматически отправляет уведомления студентам.
      tags:
        - Schedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - startTime
                - endTime
                - title
                - teacher
              properties:
                date:
                  type: string
                  format: date
                  description: Дата проведения занятия
                startTime:
                  type: string
                  pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
                  description: Время начала занятия
                endTime:
                  type: string
                  pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
                  description: Время окончания занятия
                title:
                  type: string
                  description: Тема урока
                description:
                  type: string
                  description: Описание занятия
                format:
                  type: string
                  enum: [онлайн, оффлайн, запись]
                  default: онлайн
                teacher:
                  type: string
                  description: Имя преподавателя
                materials:
                  type: array
                  items:
                    type: string
                  description: Ссылки на материалы
                streamLink:
                  type: string
                  description: Ссылка на трансляцию
                homework:
                  type: string
                  description: Домашнее задание
                homeworkDeadline:
                  type: string
                  format: date
                  description: Срок сдачи ДЗ
      responses:
        '201':
          description: Занятие успешно создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Занятие успешно создано"
                  data:
                    $ref: '#/components/schemas/Schedule'
        '400':
          description: Неверные данные
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав
        '500':
          description: Ошибка сервера
...